Update gitignore to share open scene
Students having cloned the repository would end up in a blank scene,
and need to open the correct one manually.

This is saved by Unity in Library/LastSceneManagerSetup.txt, which
was previously ignored.
Keep the rest of Library ignored, but allow this file so students
can directly open the project to the correct scene.

While we are at it, add additional ignores and remove the C# `.sln`
that is created automatically.