Unity: Use SafeSerial and provide early exits
Use the newly introduced SafeSerial wrapper, which allows us to detect
communication issues much more easily and properly abort the receiving
loop instead of getting stuck and having to restart Unity.

Update the close checks, as it should be done by the library, and take
advantage of the wrapper returning if Open() succeeded or not.