Unity: Properly reset transfer state
The current code assumes we can only completely transfer a cartridge,
so the reset occuring during the loop would be enough.

However, this is not the case anymore as we can early exit if there is
a communication issue or disconnection with the board.

Properly reset the transfer state when starting a new one.